comments and some other stuff
Basically just overhauling the orbbec module. Here I'm creating two ports for depth streams (raw and processed). The raw stream is sent directly to the Video Recorder module to be encoded in Gray16, whereas the processed stream is visualized and not generally saved. The IR stream is directly processed since it's more likely for keypoint tracking than anything else, and we don't *NEED* the raw IR pixels (tentative). Also simplified the settings dialog to only be two checkboxes (IR or depth stream?) and fixed some bugs. Hopefully this should work - but we're in dev for a reason